requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
starlette==0.38.5
pydantic==2.9.2

httpx==0.27.2
requests==2.32.3

openai==1.50.2
anthropic==0.36.0

opencv-python-headless==4.10.0.84
numpy==2.1.1

pika==1.3.2
redis==5.0.8
mysql-connector-python==9.0.0

ffmpeg-python==0.2.0
aliyun-python-sdk-core==2.16.0

python-dotenv==1.0.1

python-multipart==0.0.9
